Allow Zeros padding
This required that we either add more enum values to the WinRT-based SymmetricAlgorithm enum in a multiplying fashion, or we change to using the individual component enums for algorithm name, mode, and padding. I went with the latter.

Currently, zeros padding is only known to work on desktop framework. It fails on WinRT till I rewrite its symmetric crypto in terms of BCrypt.

Work toward #43